[fix] Fix several bugs - Now the rest backend is totally functional
* Each endpoint is totally working
* Change some endpoints related with currencies and iban
* Change maxAmount type for Double in max amount group by month
* Change some parameters to make them really optional avoiding Optional
* Add sorting phase in aggregation count per day, month and year, based on date time
* Erase some debugging logs properties related with mongo template

+++ TransferBackend/src/main/kotlin/com/ninety/nine/main/restbackend/data/data.kt
@@ -52,7 +52,7 @@ data class MonthTransferGroupCount(
 data class MonthTransferGroupMaxAmount(
     @Id
     val id: MonthAggregationKey,
-    val maxAmount: Int
+    val maxAmount: Double
 )
 
 data class MonthTransferGroupMaxCount(
